+++ game-core/src/main/java/games/strategy/engine/chat/ChatMessagePanel.java
@@ -27,7 +27,7 @@ import javax.swing.text.StyleConstants;
 import lombok.extern.java.Log;
 import org.triplea.domain.data.UserName;
 import org.triplea.java.Interruptibles;
-import org.triplea.java.TimeManager;
+import org.triplea.java.DateTimeUtil;
 import org.triplea.sound.ClipPlayer;
 import org.triplea.sound.SoundPath;
 import org.triplea.swing.SwingAction;
@@ -249,7 +249,7 @@ public class ChatMessagePanel extends JPanel implements ChatMessageListener {
 
   private void addChatMessage(final String originalMessage, final UserName from) {
     final String message = Ascii.truncate(originalMessage, 200, "...");
-    final String time = "(" + TimeManager.getLocalizedTime() + ")";
+    final String time = "(" + DateTimeUtil.getLocalizedTime() + ")";
     final Document doc = text.getDocument();
     try {
       doc.insertString(
